requirements.txt
Flask==3.0.0
pandas==2.1.4
email-validator==2.1.0
dnspython==2.4.2
rapidfuzz==3.6.1
python-dotenv==1.0.0
Werkzeug==3.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0

# Development dependencies
coverage==7.13.0
